test: consolidate screensaver interface + smoke tests into one suite
test_screensaver_interface.py looped over every screensaver and only
instantiated it; test_screensaver_smoke.py looped again to tick each one.
Merge both into test_screensavers.py so each screensaver is constructed once.

- Single setUpModule loading the real default_config.json (drops the stale
  hand-maintained mock-config dict), plus a tearDownModule resetting Config.
- Inject BlackHoleFramePlayer at construction instead of monkeypatching
  LedFramePlayer globally and never restoring it (the old interface setup
  leaked a mocked LedFramePlayer into later test modules). Injection also
  matches how ScreensaverManager builds screensavers in production.
- Fold the two instantiation tests (standard-constructor + super-init) into
  one construct-and-check pass covering all 49 screensavers, then tick the
  non-external ones. Removes a full redundant instantiation pass (49 subtests).
- Keep all metadata, video-screensaver, and preview-integration tests.

Suite still green: 76 passed, 289 subtests (was 338 — the 49 dropped subtests
were the redundant second instantiation loop).
